+++ assets/javascripts/kitten/components/cards/team-card.test.js
@@ -33,7 +33,7 @@ describe('<TeamCard />', () => {
             <TeamCard.Icons
               email="example@kisskissbankbank.com"
               phoneNumber="06 00 00 00 00"
-              socialNetworks={[
+              links={[
                 { name: 'linkedin', href: 'https://www.linkedin.com' },
                 { name: 'twitter', href: 'https://www.twitter.com' },
               ]}